tracks: Allow a smaller or non-existent border at small tile sizes
The basic tile size in Tracks is required to be a multiple of 6, which
means that for small tile sizes the steps are rather large.  With the
standard border widths, this means that the default 8-by-8 puzzle can be
246 pixels wide (tilesize == 24) or 184 pixels wide (tilesize == 18).
This is particularly annoying if you happen to have a 240-pixel-wide
screen.

This commit allows the puzzle to reduce or remove the borders at small
tile sizes, on the grounds that a slightly narrower border is acceptable
if it avoids needing to use a smaller tile size.  It encodes the border
width in (tilesize % 6), and is thus enabled when the default border
width is 5 or less.  Above that size (which is a tilesize of 48), I
assume the steps in tile size aren't big enough to be a serious problem.