js: Look up elements in the DOM as early as possible
Now that our script is loaded using <script defer>, we can rely on the
DOM's being complete as soon as it starts running.  So when we declare a
variable to point to a DOM element, we can initialise it with that
element.  This saves having these odd initialisations scattered around
the code, usually but not always at the site of first use.

I'd like to be able to do the same thing with the variables that point
to C functions, but the Module.cwrap() call isn't entirely safe before
Emscripten has finished loading the C code.