mkhighlight: Don't darken the base colour if we don't need a highlight
The "mkhighlight" functions can now take -1 as the requested highlight
or lowlight colour as an indication that no such colour should be
returned.  If the caller doesn't want a highlight colour, there's not
much point in darkening the base colour to make it distinct from the
highlight, so don't do that.

This means that the background colours in Palisade and Untangle are no
longer unnecessarily darkened.