Don't request a highlight colour in games that don't need one
This means that Range, Singles, and Tracks can now use the default
background colour even if it's close to white.  In the case of Singles
I've left a dummy entry in the colour list so as not to renumber the
rest and break everyone's environment variables.  If Singles ever needs
a new colour it can re-use that slot.