Rewrite mkhighlight to handle a wider range of colours
Before, it worked by scaling all the colour components, which worked
reasonably for colours in the vicinity of white, but failed badly on a
black background.  Now it generates the highlight and lowlight colours
by mixing the background colour with white and black respectively.
Where there isn't enough headroom, it adjusts the background by mixing
in a negative amount of white or black, which makes sense geometrically
if not in paint.

There is still a problem that green and yellow colours don't end up with
bright enough highlights.  I think this is because the code doesn't
understand that green is brighter than red or blue.