Guess: Make 'H' key work properly with "Allow duplicates" off
Before it would not only generate an invalid guess, but also override
the usual rules to allow you to submit it.  Now guesses are only
provided if they're valid, and I've adjusted the maximum-colour finder
so that the code can actually find the correct guess.  This should have
no effect on the behaviour with "Allow duplicates" turned on.