Add a dictionary for AFL++
It consists of two parts.  One is the list of all record types used by
the serialiser, to make it easy for AFL++ to find them.  The other is
the "interesting" integers used by AFL++ converted to ASCII decimal
form because Puzzles save files are coded in decimal and this will
help AFL++ to guess good values.  I hope.