Ensure we lie that the background is grey not black
Fix #585. Unfortunately the backends calculate colours from the background
colour in such a way that many things are black if the background is black.
For this reason we already tried to get specifically the day version of the
background colour to pass to the backend. For some reason this breaks on
API 28 and possibly others, so just add a second copy of the colour
to use for this purpose that doesn't have a night version.