Explicitly document that various function pointers can be NULL
There are various functions pointed to by the game structure that are
only called if certain flags are set.  Some of them were documented as
not being called when the flags were clear, but there was no explicit
statement that the pointers could simply be NULL in that case.  Now
there is.